refactor: shared device style
