+++ shared/src/styles/index.ts
@@ -1,3 +1,4 @@
 export * from './animations';
 export * from './borders';
 export * from './colors';
+export * from './device';
