--- src/styles/DebugStyles.ts
@@ -1,3 +1,0 @@
-export const debugBorder = {
-  border: '2px dashed rgba(255, 0, 0, 0.4)',
-};
